refactor(file-tree): relocate primitive under services/file/tree as parallel to FileManager.
Per file-manager-architecture.md the file module top level is src/main/services/file/.
DirectoryTreeBuilder is a top-level primitive parallel to FileManager (DB-backed
FileEntry vs in-memory tree + chokidar watcher are orthogonal resources), so it
moves alongside internal/, watcher/, utils/ instead of carving out a separate
src/main/file/ tree. search.ts + gitignore.ts are tree implementation details
(callers are only builder.ts and ipc.ts) and flatten into tree/ directly,
eliminating the src/main/file/utils/ naming collision with src/main/utils/file/.
