+++ packages/shared/file/types/tree.ts
@@ -2,7 +2,7 @@
  * Directory-tree primitive (RFC §12).
  *
  * Lives in shared because both the main-process `DirectoryTreeBuilder`
- * (`src/main/file/tree/`) and the renderer-side `useDirectoryTree` hook
+ * (`src/main/services/file/tree/`) and the renderer-side `useDirectoryTree` hook
  * (`src/renderer/src/hooks/useDirectoryTree.ts`) work against the **same
  * tree shape**. Splitting the classes into per-side mirrors made us write
  * the algorithms twice — there is no reason: every helper here is pure
