+++ src/main/services/file/tree/__tests__/registry.test.ts
