+++ src/main/services/file/tree/__tests__/search.test.ts
