+++ src/main/services/file/tree/gitignore.ts
