+++ src/main/services/file/tree/registry.ts
