+++ src/main/services/file/tree/search.ts
